EL2とEL3の実装クラスを分離
